requirements.txt
tldextract
PyYAML
Flask
